+++ src/tools/bin/cloudflare_dns.py
@@ -1,3 +1,4 @@
+#!/mnt/opt/nicola/tools/bin/python
 import json
 import os
 from collections import defaultdict